.hackathon/voicescope/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
patter
